add mappings and tests for logic processes

+++ src/eodc_openeo_bindings/map_array_processes.py
@@ -38,7 +38,7 @@ def map_mask(process):
     """
 
     param_dict = {
-        'mask': 'float',
+        'mask': 'int',
         'replacement': 'float'
     }
